src/requirements.txt
certifi==2019.11.28
click==7.1.2
colorama==0.4.3
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.0
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.34
py-lz4framed==0.14.0
PyGLM==1.2.0
pymongo==3.10.1
PyOpenGL==3.1.5
pypiwin32==223
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyserial==3.4
PyTurboJPEG==1.4.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
spinnaker-python==1.27.0.48
Werkzeug==1.0.1
win10toast==0.9
win32-setctime==1.0.1
wincertstore==0.2
